+++ vu/vmudm.h
@@ -1,12 +1,12 @@
 #include "vu.h"
 
-INLINE void do_mudm(short* VD, signed short* VS, unsigned short* VT)
+INLINE void do_mudm(short* VD, short* VS, short* VT)
 {
     long acc[N];
     register int i;
 
     for (i = 0; i < N; i++)
-        acc[i] = VS[i] * VT[i];
+        acc[i] = (signed short)(VS[i]) * (unsigned short)(VT[i]);
     for (i = 0; i < N; i++)
         ACC_H(i) = VS[i] >> 15;
     for (i = 0; i < N; i++)
